refactor app into modules
