+++ hushline/templates/base.html
@@ -22,7 +22,7 @@
                             <a href="#" class="dropbtn">@{{ session['username'] }} <img class="dropdown-icon" src="{{ url_for('static', filename='dropdown.png') }}" alt="Dropdown"></a>
                             <div class="dropdown-content">
                                 <ul>
-                                    <li><a href="{{ url_for('settings') }}">Settings</a></li>
+                                    <li><a href="{{ url_for('settings.index') }}">Settings</a></li>
                                     <li><a href="{{ url_for('logout') }}">Logout</a></li>
                                 </ul>
                             </div>
