+++ hushline/templates/confirm_disable_2fa.html
@@ -4,8 +4,8 @@
 {% block content %}
     <h2>Disable Two-Factor Authentication</h2>
     <p>Are you sure you want to disable 2FA?</p>
-    <form method="POST" action="{{ url_for('disable_2fa') }}">
+    <form method="POST" action="{{ url_for('settings.disable_2fa') }}">
         <button type="submit">Yes, Disable 2FA</button>
     </form>
-    <a href="{{ url_for('settings') }}">Cancel</a>
+    <a href="{{ url_for('settings.index') }}">Cancel</a>
 {% endblock %}
