+++ hushline/templates/verify_2fa_login.html
@@ -4,7 +4,7 @@
 {% block content %}
     <h2>Enter your 2FA Code</h2>
     <p>🔑 Enter the six-digit code from your authenticator app.</p>
-    <form method="POST" action="{{ url_for('verify_2fa_login') }}">
+    <form method="POST" action="{{ url_for('settings.verify_2fa_login') }}">
         {{ form.hidden_tag() }}
         <div>
             {{ form.verification_code.label }}
