--- wsgi.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-
-from app import app
-
-if __name__ == "__main__":
-    app.run()
